requirements.txt
sounddevice==0.4.6
whisper-ctranslate2==3.0.0
coqui-tts==0.22.0
requests==2.32.3
beautifulsoup4==4.12.3
numpy==1.26.4